--- a/t/issues/t5308-kvsdir-initial-path.py
+++ b/t/issues/t5308-kvsdir-initial-path.py
@@ -10,7 +10,7 @@ from flux.job import JobspecV1
 
 handle = flux.Flux()
 jobspec = JobspecV1.from_command(command=["/bin/true"], num_tasks=1, num_nodes=1)
-jobid = flux.job.submit(handle, jobspec, waitable=True)
+jobid = flux.job.submit(handle, jobspec)
 flux.job.wait(handle, jobid=jobid)
 jobdir = flux.job.job_kvs(handle, jobid)
 jobdir["foo"] = "bar"