--- a/t/job-manager/submit-wait.py
+++ b/t/job-manager/submit-wait.py
@@ -32,10 +32,10 @@ jobspec_fail = JobspecV1.from_command(["/bin/false"])
 jobs = []
 for i in range(njobs):
     if i < njobs / 2:
-        jobid = job.submit(h, jobspec, waitable=True)
+        jobid = job.submit(h, jobspec)
         print("submit: {} /bin/true".format(jobid))
     else:
-        jobid = job.submit(h, jobspec_fail, waitable=True)
+        jobid = job.submit(h, jobspec_fail)
         print("submit: {} /bin/false".format(jobid))
     jobs.append(jobid)