--- a/t/t2208-job-manager-wait.t
+++ b/t/t2208-job-manager-wait.t
@@ -25,36 +25,36 @@ test_expect_success "python FLUX_JOBID_ANY matches job.h" '
 	test $py_id = $c_id
 '
 
-test_expect_success "wait works on waitable job run with flux submit" '
-	JOBID=$(flux submit --flags waitable /bin/true) &&
+test_expect_success "wait works on job run with flux submit" '
+	JOBID=$(flux submit /bin/true) &&
 	flux job wait ${JOBID}
 '
 
-test_expect_success "wait works on waitable job run with flux-job" '
+test_expect_success "wait works on job run with flux-job" '
 	flux submit --dry-run /bin/true >true.jobspec &&
-	JOBID=$(flux job submit --flags=waitable true.jobspec) &&
+	JOBID=$(flux job submit true.jobspec) &&
 	flux job wait ${JOBID}
 '
 
-test_expect_success "wait works on inactive,waitable job" '
-	JOBID=$(flux submit --flags waitable /bin/true) &&
+test_expect_success "wait works on inactive job" '
+	JOBID=$(flux submit /bin/true) &&
 	flux job wait-event ${JOBID} clean &&
 	flux job wait ${JOBID}
 '
 
-test_expect_success "wait works on three waitable jobs in reverse order" '
-	JOB1=$(flux submit --flags waitable /bin/true) &&
-	JOB2=$(flux submit --flags waitable /bin/true) &&
-	JOB3=$(flux submit --flags waitable /bin/true) &&
+test_expect_success "wait works on three jobs in reverse order" '
+	JOB1=$(flux submit /bin/true) &&
+	JOB2=$(flux submit /bin/true) &&
+	JOB3=$(flux submit /bin/true) &&
 	flux job wait ${JOB3} &&
 	flux job wait ${JOB2} &&
 	flux job wait ${JOB1}
 '
 
-test_expect_success "wait FLUX_JOBID_ANY works on three waitable jobs" '
-	flux submit --flags waitable /bin/true >jobs.out &&
-	flux submit --flags waitable /bin/true >>jobs.out &&
-	flux submit --flags waitable /bin/true >>jobs.out &&
+test_expect_success "wait FLUX_JOBID_ANY works on three jobs" '
+	flux submit /bin/true >jobs.out &&
+	flux submit /bin/true >>jobs.out &&
+	flux submit /bin/true >>jobs.out &&
 	flux job wait >wait.out &&
 	flux job wait >>wait.out &&
 	flux job wait >>wait.out &&
@@ -91,14 +91,14 @@ test_expect_success "wait FLUX_JOBID_ANY exits 2 with no waitable jobs" '
 '
 
 test_expect_success "wait works when job terminated by exception" '
-	JOBID=$(flux submit --flags waitable sleep 120) &&
+	JOBID=$(flux submit sleep 120) &&
 	flux job raise --severity=0 ${JOBID} my-exception-message &&
 	test_expect_code 1 flux job wait ${JOBID} 2>exception.out &&
 	grep my-exception-message exception.out
 '
 
 test_expect_success "wait works when job tasks exit 1" '
-	JOBID=$(flux submit --flags waitable /bin/false) &&
+	JOBID=$(flux submit /bin/false) &&
 	test_expect_code 1 flux job wait ${JOBID} 2>false.out &&
 	grep exit false.out
 '
@@ -112,33 +112,33 @@ test_expect_success "wait --all works with no waitable jobs" '
 '
 
 test_expect_success "wait --all works with one job" '
-	flux submit --flags waitable /bin/true &&
+	flux submit /bin/true &&
 	flux job wait --all
 '
 
 test_expect_success "wait --all works with two jobs" '
-	flux submit --flags waitable /bin/true &&
-	flux submit --flags waitable /bin/true &&
+	flux submit /bin/true &&
+	flux submit /bin/true &&
 	flux job wait --all
 '
 
 test_expect_success "wait --all fails when first job fails" '
-	flux submit --flags waitable /bin/false &&
-	flux submit --flags waitable /bin/true &&
+	flux submit /bin/false &&
+	flux submit /bin/true &&
 	test_must_fail flux job wait --all
 '
 
 test_expect_success "wait --all fails when second job fails" '
-	flux submit --flags waitable /bin/true &&
-	flux submit --flags waitable /bin/false &&
+	flux submit /bin/true &&
+	flux submit /bin/false &&
 	test_must_fail flux job wait --all
 '
 
 
 test_expect_success "wait --all --verbose emits one line per successful job" '
-	flux submit --flags waitable /bin/true &&
-	flux submit --flags waitable /bin/true &&
-	flux submit --flags waitable /bin/false &&
+	flux submit /bin/true &&
+	flux submit /bin/true &&
+	flux submit /bin/false &&
 	test_must_fail flux job wait --all --verbose 2>verbose.err &&
 	test $(wc -l <verbose.err) -eq 3
 '
@@ -147,41 +147,28 @@ test_expect_success "wait fails on bad jobid, " '
 	test_expect_code 2 flux job wait 1
 '
 
-test_expect_success "wait fails on non-waitable, active job" '
+test_expect_success "a second wait fails on active job" '
 	JOBID=$(flux submit sleep 0.5) &&
+	flux job wait ${JOBID} &&
 	test_expect_code 2 flux job wait ${JOBID}
 '
 
-test_expect_success "wait fails on non-waitable, inactive job" '
+test_expect_success "a second wait fails on inactive job" '
 	JOBID=$(flux submit /bin/true) &&
 	flux job wait-event ${JOBID} clean &&
-	test_expect_code 2 flux job wait ${JOBID}
-'
-
-test_expect_success "a second wait fails on waitable, active job" '
-	JOBID=$(flux submit --flags waitable sleep 0.5) &&
 	flux job wait ${JOBID} &&
 	test_expect_code 2 flux job wait ${JOBID}
-'
-
-test_expect_success "a second wait fails on waitable, inactive job" '
-	JOBID=$(flux submit --flags waitable /bin/true) &&
-	flux job wait-event ${JOBID} clean &&
-	flux job wait ${JOBID} &&
-	test_expect_code 2 flux job wait ${JOBID}
-'
-
-test_expect_success "guest cannot submit job with WAITABLE flag" '
-	export FLUX_HANDLE_ROLEMASK=0x2 &&
-	test_must_fail flux submit --flags waitable /bin/true &&
-	unset FLUX_HANDLE_ROLEMASK
 '
 
 test_expect_success "guest cannot wait on a job" '
-	JOBID=$(flux submit --flags waitable /bin/true) &&
+	JOBID=$(flux submit /bin/true) &&
 	export FLUX_HANDLE_ROLEMASK=0x2 &&
 	test_expect_code 1 flux job wait ${JOBID} &&
 	unset FLUX_HANDLE_ROLEMASK
+'
+test_expect_success "wait works with deprecated waitable flag" '
+	JOBID=$(flux submit --flags waitable /bin/true) &&
+	flux job wait ${JOBID}
 '
 
 test_done