--- a/t/t2304-sched-simple-alloc-check.t
+++ b/t/t2304-sched-simple-alloc-check.t
@@ -23,7 +23,7 @@ test_expect_success 'run an alloc-bypass sleep job' '
 '
 test_expect_success 'a regular job fails with an alloc-check exception' '
 	test_expect_code 1 \
-	    run_timeout 30 flux submit --flags=waitable -vvv \
+	    run_timeout 30 flux submit -vvv \
 	    --wait-event=exception \
 	    -N1 /bin/true >bypass.jobid
 '