Problem: as described in #5738, requring jobs to be submitted with the `waitable` flag in order for them to work with `flux_job_wait(3)` is not really buying us much anymore, now that inactive jobs are kept in the job manager's memory by default.

This makes all jobs waitable by default.  Instance owner only, as before.

I marked this WIP for now pending
- documentation update
- need to make sure that unwaited-for jobs can be purged (and add a test)
- check performance/memory impact and mitigate if necessary
- consider whether wait should be made into a jobtap plugin instead of a builtin job manager service
- consider whether it is worth the effort to allow guests to wait on their own jobs