@garlick 2024-02-28
A brief pause to measure the impact of always adding unreaped jobs to the zombie hash when they transition to inactive, as opposed to only adding jobs that have the WAITABLE flag.  It seems that throughput is roughly the same.  This is a single node test out of the build tree on my desktop, after running three groups of 1000 jobs through first and letting the system quiet down:

Master (v0.59.0-77-g102eb148f):
```
ƒ(s=1,d=0,builddir) garlick@system76-pc:~/proj/flux-core$ src/test/throughput.py -n 10000
number of jobs: 10000
submit time:    3.106 s (3219.8 job/s)
script runtime: 41.173s
job runtime:    40.931s
throughput:     244.3 job/s (script: 242.9 job/s)
```

This PR
```
ƒ(s=1,d=0,builddir) garlick@system76-pc:~/proj/flux-core$ src/test/throughput.py -n 10000
number of jobs: 10000
submit time:    2.947 s (3393.0 job/s)
script runtime: 40.592s
job runtime:    40.461s
throughput:     247.2 job/s (script: 246.4 job/s)
```