@garlick 2024-02-28
My feeling at this point is that this change doesn't accomplish much, since:
- there's not really a use case for consumable wait status in the system instance
- the flag prevents some extra work from being done and thus has a useful purpose

I considered expanding the "result" to be teh same as `flux_job_result(3)` but as @grondo pointed out, if you want that you can trivially fetch it after `flux_job_wait(3)` returns.

I considered reimplementing wait as a jobtap plugin but the benefits of doing that are marginal at best.

So yeah let's close this.  Sorry for the noise.